[SideNav] Wrap SideNav in `ErrorBoundary` with retry   (#239945)
## Summary

Sidenav is an important but complex component. It is also part of the
core React tree. An error there might break the whole app.

Here is an example of such a sneaky bug
https://github.com/elastic/kibana/pull/238460

To fix this, I suggest we wrap the component in an error boundary.
Instead of showing an error, we will reset the component's state and
start over. when maxRetry hits, we show the error.