[streams] stop llm reasoning when request is aborted (#241550)
## Summary

We have several routes that triggers an llm workflow that can be long
running. If a request is closed early by the client, the process will
continue executing unnecessarily.

This change passes the request's abort signal to the relevant workflows
to prevent unneeded work in case the request is already terminated


### Testing

Trigger any ai workflow and abort the request (eg close the window). The
process should end with a `CancelledError`

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>